Fix some compiler errors

+++ learn/src/fluent_cpp/expressive_types/units/units.cc
@@ -3,21 +3,21 @@
 // NOLINTNEXTLINE(google-runtime-int)
 units::Meter operator"" _meter(unsigned long long value)
 {
-  return units::Meter{value};
+  return units::Meter{static_cast<double>(value)};
 }
 
 units::Meter operator"" _meter(long double value)
 {
-  return units::Meter{value};
+  return units::Meter{static_cast<double>(value)};
 }
 
 // NOLINTNEXTLINE(google-runtime-int)
 units::Kilometer operator"" _kilometer(unsigned long long value)
 {
-  return units::Kilometer(value);
+  return units::Kilometer(static_cast<double>(value));
 }
 
 units::Kilometer operator"" _kilometer(long double value)
 {
-  return units::Kilometer(value);
+  return units::Kilometer(static_cast<double>(value));
 }
